Updated and fixed some weird behavior with the low storage rk3 time stepping routine for particles.
Moved cases from david_* to examples/rk3/*.

Particle velocity is now output in binary files velxyz*.3D now.

"partview" bash script has been made more general so it can handle velocity binary file conversion.
See the partview files in the rk3 directory to see how to call it. It can now be called with different
filenames and also 2d and 3d (2 or 3 is the last input argument into script).

Updated makenek.inc so that particle routines are added to the .usr file if cmtparticles.usrp is not in
the directory. Note that if you want particles, you still must add line at top of .usr file: include "cmtparticles.usrp"
even if the file is in the directory.

Updated CMTDATA file so that correct TVD Rk3 arrays are transfered correctly between timesteps.