Replaced inlined rzero() in hsmg_schwarz_toext3d with calls to rzero() and rzero_acc()
This appears to correct a discrepancy in CPU vs. GPU runs the
discrepancy previously appeared in the mg_work array before and after
the call to hsmg_schwarz_toext3d

The array passed to rzero() is reshaped, and it's possible that when we
inlined rzero, we didn't account for the reshaping correctly.