[openacc] Fixed size for local_work in hsmg_tnsr1_3d_acc
Eliminates the error "variable in data clause is partially present
on the device: name=local_work", which occurs upon the second entry
into the data clause.  This error occured in the 17x17 case for high
polynomial orders.  Also brings the number of gmres iterations for
each timestep closer to the CPU version.

Tested on summitdev with lx1=8, lp=16, lelt=10000

Tested with nx1=8